build fixes

+++ sensor-raspi/sensor_install.sh
@@ -371,21 +371,25 @@ install_files() {
     echo "BUG_REPORT_URL=\"https://github.com/${IMAGE_PUBLISHER}/malcolm/issues\"" >> "$sensor_ver_file"
 
     # Setup MaxMind Geo IP info
+    mkdir -p /opt/arkime/etc
+    pushd /opt/arkime >/dev/null 2>&1
     MAXMIND_GEOIP_DB_LICENSE_KEY=""
 
     if [[ -f "$SHARED_DIR/maxmind_license.txt" ]]; then
     MAXMIND_GEOIP_DB_LICENSE_KEY="$(cat "$SHARED_DIR/maxmind_license.txt" | head -n 1)"
         if [[ ${#MAXMIND_GEOIP_DB_LICENSE_KEY} -gt 1 ]]; then
             for DB in ASN Country City; do
-            curl -s -S -L -o "/opt/arkime/GeoLite2-$DB.mmdb.tar.gz" "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-$DB&license_key=$MAXMIND_GEOIP_DB_LICENSE_KEY&suffix=tar.gz"
-            tar xvf "GeoLite2-$DB.mmdb.tar.gz" --wildcards --no-anchored '*.mmdb' --strip=1 --no-same-owner
-            rm -f "GeoLite2-$DB.mmdb.tar.gz"
-        done
+                curl -s -S -L -o "GeoLite2-$DB.mmdb.tar.gz" "https://download.maxmind.com/app/geoip_download?edition_id=GeoLite2-$DB&license_key=$MAXMIND_GEOIP_DB_LICENSE_KEY&suffix=tar.gz"
+                if [[ -f "GeoLite2-$DB.mmdb.tar.gz" ]]; then
+                    tar xvf "GeoLite2-$DB.mmdb.tar.gz" --wildcards --no-anchored '*.mmdb' --strip=1 --no-same-owner
+                    rm -f "GeoLite2-$DB.mmdb.tar.gz"
+                fi
+            done
         fi
     fi
-    curl -s -S -L -o /opt/arkime/etc/ipv4-address-space.csv "https://www.iana.org/assignments/ipv4-address-space/ipv4-address-space.csv"
-    curl -s -S -L -o /opt/arkime/etc/oui.txt "https://www.wireshark.org/download/automated/data/manuf"
-
+    curl -s -S -L -o ./etc/ipv4-address-space.csv "https://www.iana.org/assignments/ipv4-address-space/ipv4-address-space.csv"
+    curl -s -S -L -o ./etc/oui.txt "https://www.wireshark.org/download/automated/data/manuf"
+    popd >/dev/null 2>&1
 
     # Prepare Fluentbit and Beats repo GPG keys
     local apt_lists='/etc/apt/sources.list.d'
